Restructuring files, dividing into folders
