+++ dotnetcore-acquisition-library/src/Acquisition/DotnetCoreDependencyInstaller.ts
