+++ dotnetcore-acquisition-library/src/Acquisition/IAcquisitionInvoker.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
-import { IEventStream } from './EventStream';
+import { IEventStream } from '../EventStream/EventStream';
 import { IDotnetInstallationContext } from './IDotnetInstallationContext';
 import { IInstallationValidator } from './IInstallationValidator';
 import { InstallationValidator } from './InstallationValidator';
