+++ dotnetcore-acquisition-library/src/Acquisition/IAcquisitionWorkerContext.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
 import { Memento } from 'vscode';
-import { IEventStream } from './EventStream';
+import { IEventStream } from '../EventStream/EventStream';
 import { IAcquisitionInvoker } from './IAcquisitionInvoker';
 import { IInstallationValidator } from './IInstallationValidator';
 import { IVersionResolver } from './IVersionResolver';
