+++ dotnetcore-acquisition-library/src/Acquisition/IDotnetInstallationContext.ts
