+++ dotnetcore-acquisition-library/src/Acquisition/IInstallScriptAcquisitionWorker.ts
