+++ dotnetcore-acquisition-library/src/Acquisition/IVersionResolver.ts
