+++ dotnetcore-acquisition-library/src/Acquisition/ReleasesResult.ts
