+++ dotnetcore-acquisition-library/src/EventStream/EventType.ts
