+++ dotnetcore-acquisition-library/src/EventStream/IEvent.ts
