+++ dotnetcore-acquisition-library/src/EventStream/IEventStreamObserver.ts
