+++ dotnetcore-acquisition-library/src/EventStream/LoggingObserver.ts
