+++ dotnetcore-acquisition-library/src/EventStream/OutputChannelObserver.ts
