+++ dotnetcore-acquisition-library/src/EventStream/StatusBarObserver.ts
@@ -33,6 +33,7 @@ export class StatusBarObserver implements IEventStreamObserver {
 
     public dispose(): void {
         // Nothing to dispose
+
     }
 
     public setAndShowStatusBar(text: string, command: string, color?: string, tooltip?: string) {
