+++ dotnetcore-acquisition-library/src/IExtensionContext.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License. See License.txt in the project root for license information.
  * ------------------------------------------------------------------------------------------ */
-import { ITelemetryReporter } from './TelemetryObserver';
+import { ITelemetryReporter } from './EventStream/TelemetryObserver';
 
 export interface IExtensionContext {
     telemetryReporter: ITelemetryReporter | undefined;
