+++ dotnetcore-acquisition-library/src/extension.ts
@@ -7,19 +7,19 @@ import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import { AcquisitionInvoker } from './AcquisitionInvoker';
-import { DotnetCoreAcquisitionWorker } from './DotnetCoreAcquisitionWorker';
-import { DotnetCoreDependencyInstaller } from './DotnetCoreDependencyInstaller';
-import { EventStream } from './EventStream';
-import { DotnetAcquisitionMissingLinuxDependencies } from './EventStreamEvents';
-import { IEventStreamObserver } from './IEventStreamObserver';
+import { AcquisitionInvoker } from './Acquisition/AcquisitionInvoker';
+import { DotnetCoreAcquisitionWorker } from './Acquisition/DotnetCoreAcquisitionWorker';
+import { DotnetCoreDependencyInstaller } from './Acquisition/DotnetCoreDependencyInstaller';
+import { InstallationValidator } from './Acquisition/InstallationValidator';
+import { VersionResolver } from './Acquisition/VersionResolver';
+import { EventStream } from './EventStream/EventStream';
+import { DotnetAcquisitionMissingLinuxDependencies } from './EventStream/EventStreamEvents';
+import { IEventStreamObserver } from './EventStream/IEventStreamObserver';
+import { LoggingObserver } from './EventStream/LoggingObserver';
+import { OutputChannelObserver } from './EventStream/OutputChannelObserver';
+import { StatusBarObserver } from './EventStream/StatusBarObserver';
+import { TelemetryObserver } from './EventStream/TelemetryObserver';
 import { IExtensionContext } from './IExtensionContext';
-import { InstallationValidator } from './InstallationValidator';
-import { LoggingObserver } from './LoggingObserver';
-import { OutputChannelObserver } from './OutputChannelObserver';
-import { StatusBarObserver } from './StatusBarObserver';
-import { TelemetryObserver } from './TelemetryObserver';
-import { VersionResolver } from './VersionResolver';
 
 export function activate(context: vscode.ExtensionContext, parentExtensionId: string, extensionContext?: IExtensionContext) {
     const extension = vscode.extensions.getExtension(parentExtensionId);
