+++ dotnetcore-acquisition-library/src/test/unit/DotnetCoreAcquisitionWorker.test.ts
@@ -6,15 +6,15 @@ import * as chai from 'chai';
 import * as chaiAsPromised from 'chai-as-promised';
 import * as os from 'os';
 import * as path from 'path';
-import { DotnetCoreAcquisitionWorker } from '../../DotnetCoreAcquisitionWorker';
+import { DotnetCoreAcquisitionWorker } from '../../Acquisition/DotnetCoreAcquisitionWorker';
 import {
     DotnetAcquisitionCompleted,
     DotnetAcquisitionStarted,
     DotnetUninstallAllCompleted,
     DotnetUninstallAllStarted,
     TestAcquireCalled,
-} from '../../EventStreamEvents';
-import { EventType } from '../../EventType';
+} from '../../EventStream/EventStreamEvents';
+import { EventType } from '../../EventStream/EventType';
 import {
     MockEventStream,
     MockExtensionContext,
@@ -36,6 +36,7 @@ suite('DotnetCoreAcquisitionWorker Unit Tests', () => {
             storagePath: '',
             extensionState: context,
             eventStream,
+
             acquisitionInvoker: new NoInstallAcquisitionInvoker(eventStream),
             versionResolver: new MockVersionResolver(context, eventStream),
             installationValidator: new MockInstallationValidator(eventStream),
